requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
anthropic==0.7.8